バグ修正

+++ src/main/java/com/example/demo/contoller/request/SaveRaceRequest.java
@@ -29,7 +29,7 @@ public class SaveRaceRequest {
         Integer weight;
         Integer old;
         Integer frameNumber;
-        Integer handicap;
+        Float handicap;
         RaceResult raceResult;
     }
